+++ src/main/java/roomescape/controller/PageController.java
@@ -1,0 +1,23 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class PageController {
+
+    @GetMapping("/")
+    public String home() {
+        return "home";
+    }
+
+    @GetMapping("/reservation")
+    public String reservationPage() {
+        return "new-reservation";
+    }
+
+    @GetMapping("/time")
+    public String timePage() {
+        return "time";
+    }
+}
