+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,13 +1,12 @@
 package roomescape.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
 import roomescape.dto.TimeRequest;
@@ -17,7 +16,7 @@ import roomescape.service.ReservationService;
 import java.net.URI;
 import java.util.List;
 
-@Controller
+@RestController
 public class ReservationController {
 
     private final ReservationService reservationService;
@@ -26,30 +25,13 @@ public class ReservationController {
         this.reservationService = reservationService;
     }
 
-    @GetMapping("/")
-    public String home() {
-        return "home";
-    }
-
-    @GetMapping("/reservation")
-    public String reservationPage() {
-        return "reservation";
-    }
-
-    @GetMapping("/time")
-    public String reservationTimePage() {
-        return "time";
-    }
-
     @GetMapping("/reservations")
-    @ResponseBody
     public ResponseEntity<List<ReservationResponse>> getReservations() {
         List<ReservationResponse> reservations = reservationService.findAllReservations();
         return ResponseEntity.ok(reservations);
     }
 
     @PostMapping("/reservations")
-    @ResponseBody
     public ResponseEntity<ReservationResponse> createReservation(@RequestBody ReservationRequest request) {
         ReservationResponse reservation = reservationService.createReservation(request);
         return ResponseEntity
@@ -58,14 +40,12 @@ public class ReservationController {
     }
 
     @DeleteMapping("/reservations/{id}")
-    @ResponseBody
     public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
         reservationService.deleteReservation(id);
         return ResponseEntity.noContent().build();
     }
 
     @PostMapping("/times")
-    @ResponseBody
     public ResponseEntity<TimeResponse> createTime(@RequestBody TimeRequest request) {
         TimeResponse time = reservationService.createTime(request);
         return ResponseEntity
@@ -74,14 +54,12 @@ public class ReservationController {
     }
 
     @GetMapping("/times")
-    @ResponseBody
     public ResponseEntity<List<TimeResponse>> getTimes() {
         List<TimeResponse> times = reservationService.findAllTimes();
         return ResponseEntity.ok(times);
     }
 
     @DeleteMapping("/times/{id}")
-    @ResponseBody
     public ResponseEntity<Void> deleteTime(@PathVariable Long id) {
         reservationService.deleteTime(id);
         return ResponseEntity.noContent().build();
